% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{AuthZot}
\alias{AuthZot}
\title{Authenticate to a Zotero account}
\usage{
AuthZot()
}
\description{
Authenticate to a Zotero account (get keys)
}
\details{
The verification code that appears at the end of the URL after authorization in browser should be input as verification PIN
}
\examples{

credentials <- AuthZot()
}

