% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{children}
\alias{children}
\title{Get child items of a specific item}
\usage{
children(client, item_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{item_key}{Parent item key}
}
\value{
A tibble of child items
}
\description{
Get child items of a specific item
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get child items
children <- children(client, "PARENT_ITEM_KEY")
}
}
