% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{delete_collection}
\alias{delete_collection}
\title{Delete a collection from Zotero library}
\usage{
delete_collection(client, collection_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{collection_key}{Key of the collection to delete}
}
\value{
NULL on success
}
\description{
Delete a collection from Zotero library
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Delete a collection
delete_collection(client, collection_key = "COLLECTION_KEY")
}
}
