% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{download_file}
\alias{download_file}
\title{Download file attachment to disk}
\usage{
download_file(client, item_key, filename = NULL, path = NULL)
}
\arguments{
\item{client}{Zotero API client object}

\item{item_key}{Item key}

\item{filename}{Optional filename for saving}

\item{path}{Optional path for saving}
}
\value{
Path to the saved file
}
\description{
Download file attachment to disk
}
\examples{
\dontrun{
# Download file
file_path <- download_file(client, "ITEM_KEY", path = "downloads")
}
}
