% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_all_items}
\alias{get_all_items}
\title{Get all items in the library}
\usage{
get_all_items(client, chunk_size = 100)
}
\arguments{
\item{client}{Zotero API client object}

\item{chunk_size}{Number of items to retrieve per request}
}
\value{
A list of all items
}
\description{
Retrieve all items in the library by making multiple requests if necessary.
}
\examples{
\dontrun{
# Get all items
all_items <- get_all_items(client)
}
}
