% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_attachment_template}
\alias{get_attachment_template}
\title{Get attachment template}
\usage{
get_attachment_template(client, link_mode)
}
\arguments{
\item{client}{Zotero API client object}

\item{link_mode}{Link mode (imported_file, imported_url, linked_file, linked_url)}
}
\value{
Attachment template
}
\description{
Get attachment template
}
\keyword{internal}
