% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection}
\alias{get_collection}
\title{Get a specific collection from Zotero library}
\usage{
get_collection(client, collection_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{collection_key}{Collection key}
}
\value{
A list containing the specified collection
}
\description{
Get a specific collection from Zotero library
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get specific collection
collection <- get_collection(client, "COLLECTION_KEY")
}
}
