% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_collection_items_tags}
\alias{get_collection_items_tags}
\title{Get tags for items in a collection}
\usage{
get_collection_items_tags(client, collection_key, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{collection_key}{Collection key}

\item{limit}{Maximum number of tags to return}

\item{start}{Starting position}
}
\value{
A list containing collection item tags
}
\description{
Get tags for items in a collection
}
\examples{
\dontrun{
# Get collection item tags
tags <- get_collection_items_tags(client, "COLLECTION_KEY")
}
}
