% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_collection_tags}
\alias{get_collection_tags}
\title{Get tags within a specific collection}
\usage{
get_collection_tags(client, collection_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{collection_key}{Collection key}
}
\value{
A list containing collection tags
}
\description{
Get tags within a specific collection
}
\examples{
\dontrun{
# Get collection tags
tags <- get_collection_tags(client, "COLLECTION_KEY")
}
}
