% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing.R
\name{get_collection_versions}
\alias{get_collection_versions}
\title{Get versions of collections}
\usage{
get_collection_versions(client, since = NULL)
}
\arguments{
\item{client}{Zotero API client object}

\item{since}{Optional. Only return versions modified since this version}
}
\value{
A list of collection keys and their versions
}
\description{
Get a mapping of all collection keys to their current versions. This is useful
for syncing and determining which collections have changed.
}
\examples{
\dontrun{
# Get all collection versions
versions <- get_collection_versions(client)

# Get versions since a specific version
versions <- get_collection_versions(client, since = 1234)
}
}
