% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collections}
\alias{get_collections}
\title{Get collections from Zotero library}
\usage{
get_collections(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of collections to return}

\item{start}{Starting position}
}
\value{
A list containing all collections
}
\description{
Get collections from Zotero library
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get all collections
collections <- get_collections(client)
}
}
