% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collections_top}
\alias{get_collections_top}
\title{Get top-level collections from Zotero library}
\usage{
get_collections_top(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of collections to return}

\item{start}{Starting position}
}
\value{
A list containing top-level collections
}
\description{
Get top-level collections from Zotero library
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get top-level collections
top_collections <- get_collections_top(client)
}
}
