% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file}
\alias{get_file}
\title{Get file from a specific item}
\usage{
get_file(client, item_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{item_key}{Item key}
}
\value{
Raw file content
}
\description{
Get file from a specific item
}
\examples{
\dontrun{
# Get file content
file_content <- get_file(client, "ITEM_KEY")
}
}
