% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_groups}
\alias{get_groups}
\title{Get user groups}
\usage{
get_groups(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of groups to return}

\item{start}{Starting position}
}
\value{
A list of groups
}
\description{
Get all groups that the current API key has access to.
}
\examples{
\dontrun{
# Get all accessible groups
groups <- get_groups(client)
}
}
