% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_item_subset}
\alias{get_item_subset}
\title{Get a subset of items by their keys}
\usage{
get_item_subset(client, item_keys)
}
\arguments{
\item{client}{Zotero API client object}

\item{item_keys}{Vector of item keys}
}
\value{
A list of items
}
\description{
Retrieve multiple items in a single request by providing their keys.
}
\examples{
\dontrun{
# Get specific items
items <- get_item_subset(client, c("KEY1", "KEY2", "KEY3"))
}
}
