% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_item_tags}
\alias{get_item_tags}
\title{Get tags for a specific item}
\usage{
get_item_tags(client, item_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{item_key}{Item key}
}
\value{
A list containing item tags
}
\description{
Get tags for a specific item
}
\examples{
\dontrun{
# Get item tags
tags <- get_item_tags(client, "ITEM_KEY")
}
}
