% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing.R
\name{get_item_versions}
\alias{get_item_versions}
\title{Get versions of items}
\usage{
get_item_versions(client, since = NULL)
}
\arguments{
\item{client}{Zotero API client object}

\item{since}{Optional. Only return versions modified since this version}
}
\value{
A list of item keys and their versions
}
\description{
Get a mapping of all item keys to their current versions. This is useful for
syncing and determining which items have changed.
}
\examples{
\dontrun{
# Get all item versions
versions <- get_item_versions(client)

# Get versions since a specific version
versions <- get_item_versions(client, since = 1234)
}
}
