% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_items_tags}
\alias{get_items_tags}
\title{Get all item tags with filtering options}
\usage{
get_items_tags(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of tags to return}

\item{start}{Starting position}
}
\value{
A list containing filtered tags
}
\description{
Get all item tags with filtering options
}
\examples{
\dontrun{
# Get all item tags
tags <- get_items_tags(client)
}
}
