% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_items_top_tags}
\alias{get_items_top_tags}
\title{Get tags for top-level items}
\usage{
get_items_top_tags(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of tags to return}

\item{start}{Starting position}
}
\value{
A list containing top-level item tags
}
\description{
Get tags for top-level items
}
\examples{
\dontrun{
# Get top-level item tags
tags <- get_items_top_tags(client)
}
}
