% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_items_trash_tags}
\alias{get_items_trash_tags}
\title{Get tags for items in trash}
\usage{
get_items_trash_tags(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of tags to return}

\item{start}{Starting position}
}
\value{
A list containing trash item tags
}
\description{
Get tags for items in trash
}
\examples{
\dontrun{
# Get trash item tags
tags <- get_items_trash_tags(client)
}
}
