% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing.R
\name{get_last_modified_version}
\alias{get_last_modified_version}
\title{Get the last modified version of the library}
\usage{
get_last_modified_version(client)
}
\arguments{
\item{client}{Zotero API client object}
}
\value{
Integer representing the last modified version
}
\description{
Get the version number of the last change to the library. This can be used
as a starting point for syncing.
}
\examples{
\dontrun{
# Get last modified version
version <- get_last_modified_version(client)
}
}
