% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{get_publications}
\alias{get_publications}
\title{Get items from My Publications}
\usage{
get_publications(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of items to return}

\item{start}{Starting position}
}
\value{
A list of publication items
}
\description{
Get items from the user's My Publications collection.
}
\examples{
\dontrun{
# Get publications
publications <- get_publications(client)
}
}
