% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_publications_tags}
\alias{get_publications_tags}
\title{Get tags for items in My Publications}
\usage{
get_publications_tags(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of tags to return}

\item{start}{Starting position}
}
\value{
A list containing publication item tags
}
\description{
Get tags for items in My Publications
}
\examples{
\dontrun{
# Get publication item tags
tags <- get_publications_tags(client)
}
}
