% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{get_search}
\alias{get_search}
\title{Get a specific saved search from Zotero library}
\usage{
get_search(client, search_key)
}
\arguments{
\item{client}{Zotero API client object}

\item{search_key}{Search key}
}
\value{
A list containing the specified saved search
}
\description{
Get a specific saved search from Zotero library
}
\examples{
\dontrun{
client <- zotero_client(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get specific saved search
search <- get_search(client, "SEARCH_KEY")
}
}
