% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_subcollections}
\alias{get_subcollections}
\title{Get subcollections of a specific collection}
\usage{
get_subcollections(client, collection_key, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{collection_key}{Parent collection key}

\item{limit}{Maximum number of subcollections to return}

\item{start}{Starting position}
}
\value{
A list containing subcollections
}
\description{
Get subcollections of a specific collection
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get subcollections
subcollections <- get_subcollections(client, "PARENT_COLLECTION_KEY")
}
}
