% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags}
\alias{get_tags}
\title{Get all tags from Zotero library}
\usage{
get_tags(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of tags to return}

\item{start}{Starting position}
}
\value{
A list containing all tags
}
\description{
Get all tags from Zotero library
}
\examples{
\dontrun{
client <- zotero_client(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get all tags
tags <- get_tags(client)
}
}
