% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-zotero-data.R
\name{get_zotero_data}
\alias{get_zotero_data}
\title{Retrieve data from Zotero database}
\usage{
get_zotero_data(path = "~/Documents/Zotero")
}
\arguments{
\item{path}{a \code{character} of length 1. The folder containing the Zotero
database named \code{zotero.sqlite}.}
}
\value{
A \code{data.frame} with references in row and the following variables:
\itemize{
\item \code{item_id}: the identifier of the publication in the Zotero database;
\item \code{library_type}: the type of the library. Can be \code{user} (My library)
or \code{group} (Group libraries);
\item \code{group}: the name of the group library;
\item \code{collection}: the name of the library collection;
\item \code{category}: the category of the publication (\code{article}, \code{book},
\code{book_section}, \code{report}, etc.);
\item \code{year}: the year of the publication;
\item \code{title}: the title of the publication;
\item \code{author}: the authors of the publication;
\item \code{journal}: the journal of the publication (only for \code{article});
\item \code{book_title}: the title of the book (only for \code{book_section});
\item \code{editor}: the editors of the book (only for \code{book} and \code{book_section});
\item \code{abstract}: the abstract of the publication;
\item \code{volume}: the volume of the publication (only for \code{article});
\item \code{issue}: the issue of the publication (only for \code{article});
\item \code{pages}: the pages of the publication;
\item \code{publisher}: the publisher name;
\item \code{place}: the location of the publisher;
\item \code{institution}: the institution name;
\item \code{doi}: the DOI (Digital Object Identifier) of the publication;
\item \code{url}: the URL of the publication;
\item \code{note}: the annotations associated to the publication (only the \code{PUBID}
is retrieved).
}
}
\description{
Retrieves references metadata from Zotero local database (SQLite).
}
\examples{
\dontrun{
references <- get_zotero_data()
}
}
