% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{make_request}
\alias{make_request}
\title{Make a request to Zotero API}
\usage{
make_request(
  client,
  endpoint,
  method = "GET",
  query = list(),
  body = NULL,
  format = "json",
  ...
)
}
\arguments{
\item{client}{Zotero API client object}

\item{endpoint}{API endpoint}

\item{method}{HTTP method}

\item{query}{Query parameters}

\item{body}{Request body}

\item{format}{Response format, defaults to "json"}
}
\description{
Make a request to Zotero API
}
\keyword{internal}
