% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_citations.R
\name{pick_citations}
\alias{pick_citations}
\title{A Function to pick citations from Zotero}
\usage{
pick_citations()
}
\value{
Adds reference keys in your markdown document
}
\description{
This function allows you to select citations from your zotero library to add the appropriate keys to your markdown document
}
\examples{
pick_citations()

}

