% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-demo.R
\name{print_msg}
\alias{print_msg}
\title{Print a message}
\usage{
print_msg(x = "Hello world")
}
\arguments{
\item{x}{a character of length 1. Default is \code{'Hello world'}.}
}
\value{
The value of \code{x} (only if the result is assigned to a variable).
}
\description{
This function prints a simple message. This is a demo function to show good
practices in writing and documenting R function. If you delete this function
do not forget to delete the corresponding test file
\code{tests/testthat/test-demo.R} if you used \code{new_package(test = TRUE)}.
}
\examples{
print_msg()
print_msg("Bonjour le monde")

x <- print_msg("Bonjour le monde")
x
}
