% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadZotItem}
\alias{ReadZotItem}
\title{Read details of a Zotero item}
\usage{
ReadZotItem(user, id, credentials)
}
\arguments{
\item{user}{UserId}
}
\description{
Read details of a Zotero item
}
\examples{

item <- ReadZotItem()
}

