% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{trash}
\alias{trash}
\title{Get items in trash}
\usage{
trash(client, limit = 10, start = 0)
}
\arguments{
\item{client}{Zotero API client object}

\item{limit}{Maximum number of items to return}

\item{start}{Starting position}
}
\value{
A tibble of items in trash
}
\description{
Get items in trash
}
\examples{
\dontrun{
client <- zotero_api(
  library_id = "123456",
  library_type = "user",
  api_key = "your_api_key"
)

# Get items in trash
trash_items <- trash(client)
}
}
