% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing.R
\name{update_item_version}
\alias{update_item_version}
\title{Update an item with version check}
\usage{
update_item_version(client, item_key, data, version)
}
\arguments{
\item{client}{Zotero API client object}

\item{item_key}{Item key to update}

\item{data}{New item data}

\item{version}{Current version of the item}
}
\value{
Updated item data
}
\description{
Update an item while ensuring no conflicts with other changes. If the item
has been modified since the specified version, the update will fail.
}
\examples{
\dontrun{
# Update an item
result <- update_item_version(client, 
                             item_key = "ITEM_KEY",
                             data = list(title = "New Title"),
                             version = 123)
}
}
