% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing.R
\name{update_items_version}
\alias{update_items_version}
\title{Update multiple items with version check}
\usage{
update_items_version(client, items, version)
}
\arguments{
\item{client}{Zotero API client object}

\item{items}{List of items to update}

\item{version}{Current library version}
}
\value{
Update response
}
\description{
Update multiple items while ensuring no conflicts with other changes. If any
item has been modified since the specified version, the entire update will fail.
}
\examples{
\dontrun{
# Update multiple items
items <- list(
  list(key = "KEY1", data = list(title = "New Title 1")),
  list(key = "KEY2", data = list(title = "New Title 2"))
)
result <- update_items_version(client, items, version = 123)
}
}
