% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{Upload file to authorized URL}
\usage{
upload_file(url, file_path, params)
}
\arguments{
\item{url}{Upload URL}

\item{file_path}{Path to file}

\item{params}{Upload parameters}
}
\value{
Response from upload
}
\description{
Upload file to authorized URL
}
\keyword{internal}
