% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zot_convert_creators_to_df}
\alias{zot_convert_creators_to_df}
\title{Transform a characther vectors of creators (authors, etc.) into a list of properly categorised data frames}
\usage{
zot_convert_creators_to_df(creator)
}
\arguments{
\item{item_type}{Defaults to "book". It must correspond to a valid item type. You can chech which item types are valid with the function `zot_get_item_types()`}

\item{cache}{Logical, defaults to TRUE. If TRUE, it stores the template in a "zot_cache" folder in the current working directory.}
}
\value{
A data frame, with one column for each accepted input for the given item type.
}
\description{
Transform a characther vectors of creators (authors, etc.) into a list of properly categorised data frames
}
\examples{

zot_convert_creators_to_df(creator = "De Gasperi, Alcide")
}
