% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zot_convert_creators_to_df_list}
\alias{zot_convert_creators_to_df_list}
\title{Transform a characther vectors of creators (authors, etc.) into a list of properly categorised data frames}
\usage{
zot_convert_creators_to_df_list(
  creator,
  creator_type = "author",
  separator_multiple_authors = "; ",
  separator_surname_name = ", "
)
}
\arguments{
\item{creator}{A character vector (usually, a column in a data frame). Names must be given in the format: "Surname, Name; Surname, Name" (use semicolumn only to separate between multiple authors).}

\item{creator_type}{Defaults to "author". Must be a character vector of length 1.}
}
\value{
A data frame, with one column for each accepted input for the given item type.
}
\description{
Transform a characther vectors of creators (authors, etc.) into a list of properly categorised data frames
}
\examples{

zot_convert_creators_to_df_list(creator = "Spinelli, Altiero; Rossi, Ernesto")
}
