% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_item.R
\name{zot_create_items}
\alias{zot_create_items}
\title{Create new items from a data frame}
\usage{
zot_create_items(item_df, collection = NULL, user = NULL, credentials = NULL)
}
\arguments{
\item{item_df}{A data frame representing new items. Column names must correspond to `zoteroR::zot_get_item_template()` for the given item type.}

\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
Nothing, used for its side effects (creates new items on Zotero)
}
\description{
Create new Zotero items from a data frame
}
\examples{

item <- zot_create_items(item_df)
}
