% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{zot_get_item_template}
\alias{zot_get_item_template}
\title{Get an item template for a valid item type}
\usage{
zot_get_item_template(item_type = "book", cache = TRUE)
}
\arguments{
\item{item_type}{Defaults to "book". It must correspond to a valid item type. You can chech which item types are valid with the function `zot_get_item_types()`}

\item{cache}{Logical, defaults to TRUE. If TRUE, it stores the template in a "zot_cache" folder in the current working directory.}
}
\value{
A charachter vector including all categories in which given item is included.
}
\description{
Get an item template for a valid item type
}
\examples{

item_types <- zot_get_item_types()
}
