% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{zot_read_children}
\alias{zot_read_children}
\title{Read details of all children of a Zotero item}
\usage{
zot_read_children(id, user = NULL, credentials = NULL)
}
\arguments{
\item{id}{Id code of a zotero item}

\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
A data.frame including details on all children of a given parent item.
}
\description{
Read details of all children of a Zotero item
}
\examples{

item <- ZotReadChildren()
}
