% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{zot_read_children_id}
\alias{zot_read_children_id}
\title{Read IDs of all children of a Zotero item}
\usage{
zot_read_children_id(id, user = NULL, credentials = NULL)
}
\arguments{
\item{id}{Id code of a zotero item}

\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
A vector including the ID of all children
}
\description{
Read IDs of all children of a Zotero item
}
\examples{

item_id <- zot_read_children_id()
}
