% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{zot_set_options}
\alias{zot_set_options}
\title{Sets 'zoteroR' options}
\usage{
zot_set_options(...)
}
\arguments{
\item{user}{A Zotero userId.}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or
an API secret key with write access created at
[https://www.zotero.org/settings/keys]}
}
\value{
Nothing. Used for its side effects (stores settings).
}
\description{
It allows to preliminary store options frequently used by 'zoteroR', thus
removing the requirement to specify them each time a function is called.
}
\examples{
zot_set_options(user = 12345, credentials = "<API>")
}
