% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{zot_which_collection}
\alias{zot_which_collection}
\title{Extract collections in which given item is included}
\usage{
zot_which_collection(id, user = NULL, credentials = NULL)
}
\arguments{
\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
A charachter vector including all categories in which given item is included.
}
\description{
Extract collections in which given item is included
}
\examples{
\dontrun{
categories <- zot_which_collection(item = "X1X2X3")
}
}
