% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editItem.R
\name{ZotAddToCollection}
\alias{ZotAddToCollection}
\title{Add item to a collection}
\usage{
ZotAddToCollection(id, collectionId, user = NULL, credentials = NULL)
}
\arguments{
\item{id}{Id code of a zotero item}

\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
A list including all available details on a given item.
}
\description{
Add item to a collection
}
\examples{

item <- ZotAddToCollection(id = "<itemId>", collectionId = "<collectionId>")
}
