% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCollection.R
\name{ZotCreateCollection}
\alias{ZotCreateCollection}
\title{Create a collection in Zotero}
\usage{
ZotCreateCollection(collectionName, user = NULL, credentials = NULL)
}
\arguments{
\item{collectionName}{Name of the collection to be added}

\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
The key of the newly created collection (or of the pre-existing collection, if already one with the same name exists) as a character vector
}
\description{
Create a collection in Zotero. If a collection by the same name exists, it does not create a new one, but rather outputs the id of that collection.
}
\details{
Creating a collection needs an API with write access
}
\examples{

key <- ZotCreateCollection(user = 12345, collectionName = "ZoteroRtest", credentials = "<API>")
}
