% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoteror-package.R
\docType{package}
\name{zoteror-package}
\alias{zoteror}
\alias{zoteror-package}
\title{zoteror: Handle Zotero Local Database}
\description{
Makes a connection to the (local) Zotero database (SQLite) to extract bibliographic references metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/zoteror}
  \item Report bugs at \url{https://github.com/FRBCesab/zoteror/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
