% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{ZotSize}
\alias{ZotSize}
\title{Find size of zotero objects stored locally}
\usage{
ZotSize(path)
}
\arguments{
\item{path}{Path to local Zotero storage sub-folde, e.g. "/home/user/.mozilla/firefox/mdk7yirc.default/zotero/storage"}
}
\value{
A data.frame (tibble) with ID and size of each item present in the local storage.
}
\description{
Find size of zotero objects stored locally
}
\examples{

size <- ZotSize()
}
