% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ZotWhichCollection}
\alias{ZotWhichCollection}
\title{Extract categories in which given item is included}
\usage{
ZotWhichCollection(id, user = NULL, credentials = NULL)
}
\arguments{
\item{user}{Zotero userId}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
A charachter vector including all categories in which given item is included.
}
\description{
Extract categories in which given item is included
}
\examples{

categories <- ZotWhichCollection(item = "X1X2X3")
}

