% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{AuthZot}
\alias{AuthZot}
\title{Authenticate to a Zotero account}
\usage{
AuthZot(store = FALSE)
}
\arguments{
\item{store}{Logical, defaults to FALSE. If TRUE, it stores the credentials in the working diretory in a file called "ZoteroCredentials.rds", which should be considered confidential since it provides access to a given Zotero account.}
}
\description{
Authenticate to a Zotero account (get keys)
}
\details{
The verification code that appears at the end of the URL after authorization in browser should be input as verification PIN. If the parametere store is enabled - AuthZot(store=TRUE) - zoteroR stores the credentials in a local file called "ZoteroCredentials.rds", which should be considered confidential since it provides access to a given Zotero account.
If a pre-existing "ZoteroCredentials.rds" exists, it is loaded automatically.
}
\examples{

credentials <- AuthZot(store=TRUE)
}

