% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bibtex.R
\name{create_bibtex}
\alias{create_bibtex}
\title{A Function to create a bibtex file}
\usage{
create_bibtex()
}
\description{
This function allows you to create a bibtex file for the citations used in your markdown document.
}
\examples{
create_bibtex()
}

