% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCollection.R
\name{CreateZotCollection}
\alias{CreateZotCollection}
\title{Create a collection in Zotero}
\usage{
CreateZotCollection(user, collectionName, credentials)
}
\arguments{
\item{user}{Zotero userId}

\item{collectionName}{Name of the collection to be added}

\item{credentials}{Either an R object created with AuthZot(store = TRUE), or an API secret key with write access created at https://www.zotero.org/settings/keys}
}
\value{
The key of the newly created collection (or of the pre-existing collection, if already one with the same name exists) as a character vector
}
\description{
Create a collection in Zotero
}
\details{
Creating a collection needs an API with write access
}
\examples{

CreateZotCollection(user = 12345, collectionName = "ZoteroRtest", credentials = "<API>")
}

