% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ZotReadItem}
\alias{ZotReadItem}
\title{Read details of a Zotero item}
\usage{
ZotReadItem(user, id, credentials)
}
\arguments{
\item{user}{Zotero userId}
}
\value{
A list including all available details on a given item.
}
\description{
Read details of a Zotero item
}
\examples{

item <- ZotReadItem()
}

